# internal/apidocs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "go-template 的后端 API，包含安装向导、认证、管理员用户管理与系统设置接口。",
        "title": "go-template API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "修改当前用户密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_user.ChangePasswordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_httpserver.changePasswordResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录凭据",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpserver.loginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_httpserver.loginResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    }
                }
            }
        },
...
        "/api/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "获取当前登录用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_httpserver.currentUserResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "刷新访问令牌",
                "parameters": [
                    {
                        "description": "刷新令牌",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpserver.refreshPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_httpserver.refreshResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    }
                }
            }
        },
        "/api/setup/install": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup"
                ],
                "summary": "执行首次安装",
                "parameters": [
                    {
                        "description": "安装向导配置",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_setup.InstallInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_httpserver.installResultData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    }
                }
            }
        },
...
        "/api/system/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "获取系统设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_httpserver.settingsResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ysicing_go-template_internal_shared.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ysicing_go-template_internal_auth.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_ysicing_go-template_internal_config.CacheConfig": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                },
                "db": {
                    "type": "integer"
                },
                "driver": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_ysicing_go-template_internal_config.DatabaseConfig": {
            "type": "object",
            "properties": {
                "driver": {
                    "type": "string"
                },
                "dsn": {
                    "type": "string"
                }
            }
        },
        "github_com_ysicing_go-template_internal_config.JWTConfig": {
            "type": "object",
            "properties": {
                "access_ttl": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "refresh_ttl": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "github_com_ysicing_go-template_internal_config.LogConfig": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                }
            }
        },
        "github_com_ysicing_go-template_internal_config.ServerConfig": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
        "github_com_ysicing_go-template_internal_setup.InstallInput": {
            "type": "object",
            "properties": {
                "admin_email": {
                    "type": "string"
                },
                "admin_password": {
                    "type": "string"
                },
                "admin_username": {
                    "type": "string"
                },
                "cache": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_config.CacheConfig"
                },
                "database": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_config.DatabaseConfig"
                },
                "jwt": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_config.JWTConfig"
                },
                "log": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_config.LogConfig"
                },
                "server": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_config.ServerConfig"
                }
            }
        },
        "github_com_ysicing_go-template_internal_shared.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_ysicing_go-template_internal_system.Setting": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_ysicing_go-template_internal_user.ChangePasswordInput": {
            "type": "object",
            "properties": {
                "confirm_new_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ysicing_go-template_internal_user.Role": {
            "type": "string",
            "enum": [
                "user",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin"
            ]
        },
...
        "github_com_ysicing_go-template_internal_user.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_at": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_user.Role"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_httpserver.changePasswordResponseData": {
            "type": "object",
            "properties": {
                "changed": {
                    "type": "boolean"
                }
            }
        },
        "internal_httpserver.currentUserResponseData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_user.User"
                }
            }
        },
...
        "internal_httpserver.installResultData": {
            "type": "object",
            "properties": {
                "installed": {
                    "type": "boolean"
                }
            }
        },
        "internal_httpserver.loginPayload": {
            "type": "object",
            "properties": {
                "identifier": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_httpserver.loginResponseData": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_auth.TokenPair"
                },
                "user": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_user.User"
                }
            }
        },
...
        "internal_httpserver.refreshPayload": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "internal_httpserver.refreshResponseData": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/github_com_ysicing_go-template_internal_auth.TokenPair"
                }
            }
        },
        "internal_httpserver.settingsResponseData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ysicing_go-template_internal_system.Setting"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer access token，例如：Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}